tools/marketing/pyproject.toml
[tool.setuptools]
py-modules = []

[project]
name = "babel-studio-marketing"
version = "0.1.0"
description = "Developer marketing tooling — GitHub social previews, screenshots, brand assets"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "Pillow>=11.0.0",
    "playwright>=1.49.0",
    "jinja2>=3.1.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
]

[tool.ruff]
line-length = 100
